Don't allow removing files from the tagger twice. Fixes ticket #2966 and at least parts of #4822

+++ picard/tagger.py
@@ -436,9 +436,10 @@ class Tagger(QtGui.QApplication):
     def remove_files(self, files, from_parent=True):
         """Remove files from the tagger."""
         for file in files:
-            self.analyze_queue.remove(file.filename)
-            del self.files[file.filename]
-            file.remove(from_parent)
+            if self.files.has_key(file.filename):
+                self.analyze_queue.remove(file.filename)
+                del self.files[file.filename]
+                file.remove(from_parent)
 
     def remove_album(self, album):
         """Remove the specified album."""
